refactor: move validate command under tdp vars
