+++ tdp/cli/commands/vars/validate.py
